Move tests from validator_rules_test.js back to validator_test.js
