+++ validator/build.py
@@ -63,9 +63,9 @@ def CheckPrereqs():
   for f in [
       'validator-main.protoascii', 'validator.proto', 'validator_gen_js.py',
       'package.json', 'engine/validator.js', 'engine/validator_test.js',
-      'engine/validator_rules_test.js', 'engine/validator-in-browser.js',
-      'engine/tokenize-css.js', 'engine/definitions.js', 'engine/parse-css.js',
-      'engine/parse-srcset.js', 'engine/parse-url.js'
+      'engine/validator-in-browser.js', 'engine/tokenize-css.js',
+      'engine/definitions.js', 'engine/parse-css.js', 'engine/parse-srcset.js',
+      'engine/parse-url.js'
   ]:
     if not os.path.exists(f):
       Die('%s not found. Must run in amp_validator source directory.' % f)
@@ -471,32 +471,6 @@ def CompileValidatorTestMinified(out_dir):
   logging.info('... success')
 
 
-def CompileValidatorRulesTestMinified(out_dir):
-  """Runs closure compiler for validator_rules_test.js.
-
-  Args:
-    out_dir: directory name of the output directory. Must not have slashes,
-      dots, etc.
-  """
-  logging.info('entering ...')
-  CompileWithClosure(
-      js_files=[
-          'engine/definitions.js', 'engine/htmlparser.js',
-          'engine/parse-css.js', 'engine/parse-srcset.js',
-          'engine/parse-url.js', 'engine/tokenize-css.js',
-          '%s/validator-generated.js' % out_dir,
-          '%s/validator-proto-generated.js' % out_dir,
-          'engine/validator-in-browser.js', 'engine/validator.js',
-          'engine/amp4ads-parse-css.js', 'engine/keyframes-parse-css.js',
-          'engine/htmlparser-interface.js', 'light/dom-walker.js',
-          'engine/validator_rules_test.js'
-      ],
-      definitions=[],
-      entry_points=['amp.validator.ValidatorRulesTest'],
-      output_file='%s/validator_rules_test_minified.js' % out_dir)
-  logging.info('... success')
-
-
 def CompileValidatorLightTestMinified(out_dir):
   """Runs closure compiler for validator-light_test.js.
 
@@ -675,7 +649,6 @@ def GenerateTestRunner(out_dir):
              var jasmine = new JasmineRunner();
              process.env.TESTDATA_ROOTS = 'testdata:%s'
              require('./validator_test_minified');
-             require('./validator_rules_test_minified');
              require('./validator-light_test_minified');
              require('./htmlparser_test_minified');
              require('./parse-css_test_minified');
@@ -728,7 +701,6 @@ def Main(parsed_args):
   RunSmokeTest(out_dir='dist')
   RunIndexTest()
   CompileValidatorTestMinified(out_dir='dist')
-  CompileValidatorRulesTestMinified(out_dir='dist')
   CompileValidatorLightTestMinified(out_dir='dist')
   CompileHtmlparserTestMinified(out_dir='dist')
   CompileParseCssTestMinified(out_dir='dist')
